move response api out of spec
