+++ sgl-router/src/protocols/mod.rs
@@ -7,6 +7,7 @@ pub mod completion;
 pub mod embedding;
 pub mod generate;
 pub mod rerank;
+pub mod responses;
 pub mod sampling_params;
 pub mod validated;
 pub mod worker_spec;
@@ -22,6 +23,7 @@ pub use completion::*;
 pub use embedding::*;
 pub use generate::*;
 pub use rerank::*;
+pub use responses::*;
 pub use sampling_params::*;
 pub use validated::*;
 pub use worker_spec::*;
